Rename files and change their extensions to .jsx
