+++ app/components/views/AccountsPage/Accounts/AccountRow/helpers/AccountDetails.jsx
