+++ app/components/views/AccountsPage/Accounts/AccountRow/helpers/RenameAccount.jsx
