+++ app/components/views/AccountsPage/Privacy/Privacy.module.css
@@ -2,16 +2,13 @@
   display: flex;
   flex-direction: row;
 }
-
 .isColumn {
   display: flex;
   flex-direction: column;
 }
-
 .bold {
   font-weight: 700;
 }
-
 .accountsSelect {
   width: 300px;
 }
