+++ app/components/views/AccountsPage/Privacy/PrivacyContent.jsx
@@ -6,7 +6,7 @@ import "style/Privacy.less";
 import { classNames } from "pi-ui";
 import style from "./Privacy.module.css";
 
-const PrivacyPage = ({
+const PrivacyContent = ({
   mixedAccountName,
   mixedAccountBranch,
   changeAccountName,
@@ -19,7 +19,6 @@ const PrivacyPage = ({
 }) => (
     <>
       <Subtitle title={<T id="privacy.subtitle" m="Privacy" />} />
-
       <div className={classNames(style.privacyPageWrapper, "is-column")}>
         <div className={classNames(style.isRow, style.privacyRow)}>
           <div className={classNames(style.isRow, style.privacyItem)}>
@@ -28,7 +27,6 @@ const PrivacyPage = ({
           </div>
             <TextInput required disabled value={mixedAccountName} />
           </div>
-
           <div className={style.privacyItem}>
             <div className={style.isRow}>
               <div className={""}>
@@ -38,7 +36,6 @@ const PrivacyPage = ({
             </div>
           </div>
         </div>
-
         <div className={classNames(style.isRow, style.privacyRow)}>
           <div className={classNames(style.isRow, style.privacyItem)}>
             <div className={""}>
@@ -47,7 +44,6 @@ const PrivacyPage = ({
             <TextInput required disabled value={changeAccountName} />
           </div>
         </div>
-
         <div className={classNames(style.isRow, style.privacyRow)}>
           <div className={classNames(style.isRow, style.privacyItem)}>
             <div className={""}>
@@ -55,7 +51,6 @@ const PrivacyPage = ({
           </div>
             <TextInput required disabled value={csppServer} />
           </div>
-
           <div className={classNames(style.isRow, style.privacyItem)}>
             <div className={""}>
               <T id="privacy.mixing.server.port" m="Shuffle Port" />:
@@ -91,4 +86,4 @@ const PrivacyPage = ({
     </>
   );
 
-export default PrivacyPage;
+export default PrivacyContent;
